Moved Pokemon Type preprocessing unit to data_utils.py
